refactor(app): trim feature aggregation plumbing
